move lib code from main to binding
