+++ sgl-router/bindings/python/Cargo.toml
@@ -9,10 +9,12 @@ crate-type = ["cdylib"]
 
 [dependencies]
 pyo3 = { version = "0.27.1", features = ["extension-module", "abi3-py38"] }
+tokio = { version = "1.42.0", features = ["full"] }
 
 [dependencies.sglang-router]
 path = "../.."
 default-features = true
+package = "sglang-router"
 
 [features]
 default = ["pyo3/extension-module"]
